requirements.txt
azure-storage-blob
azure-ai-documentIntelligence
streamlit
python-dotenv
azure.core 
